Better command parsing, hide internal roles
We now parse the command string rather carefully in comparison to
previous rough efforts: flags are checked to see whether they are
switches before further words are automatically associated with them,
and switches and flags belonging to commands which are not the ultimate
subcommand are still tagged.

Roles which have a kind beginning with @ are now hidden as internal.