Decide to publish based on git tag (#199)
This change will make publishing as simple as ensuring there's a tag
on the commit corresponding to the declared version. Unfortunately
most of the time this will mean the job will have to be restarted
after pushing the tag. However at the price of this minor
inconvenience releasing is simplified and tied to git workflow.